fuck_the_exam/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
python-dotenv
pymupdf
faiss-cpu
numpy
sentence-transformers
pytesseract
Pillow

